refactored app.py into a subfolder dividng functionality
